requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-dotenv
pytest
pytest-asyncio
# LLM Providers
openai>=1.12.0
anthropic>=0.18.0
tiktoken>=0.5.2

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# Data Validation & Processing
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Async & Concurrency
asyncio>=3.4.3

# Database (for future memory/storage)
# sqlalchemy>=2.0.25
# alembic>=1.13.1

# Monitoring & Observability
structlog>=24.1.0
prometheus-client>=0.19.0
opentelemetry-distro>=0.45b0
opentelemetry-instrumentation-fastapi>=0.45b0
opentelemetry-exporter-otlp-proto-grpc>=1.24.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0  # For testing async

# Development Tools
black>=24.1.0
ruff>=0.1.14
mypy>=1.8.0
pre-commit>=3.6.0

# Vector Store (for RAG)
chromadb>=0.4.22

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0

# Optional: Shopify Integration
# shopify-python-api>=12.4.0

# Optional: Notification Services
# slack-sdk>=3.26.0
# sendgrid>=6.11.0